Fixed bug in AddEdge (no longer leaving stumps); updated LayerTests so that nbr edges now described in ulong (not uint).

+++ Threadle.Tests/LayerTests.cs
@@ -750,7 +750,7 @@ public class LayerTests
         net.AddEdge("friends", 2, 3);
         net.AddEdge("friends", 4, 5);
 
-        uint edgesBefore = ((ILayerOneMode)net.Layers["friends"]).NbrEdges;
+        ulong edgesBefore = ((ILayerOneMode)net.Layers["friends"]).NbrEdges;
         net.Pack("friends");
         net.Unpack("friends");
